Unit Creation: allocate_fare.cpp
Changelog//
+ Addition of the allocate_fare class
+ Implementations into main.cpp (the main menu)
+ Addition of 'status' row to both vehicles and drivers, this will be used by the operator to determine whether they are active or inactive.

* Renamed buttons.tsv to config.tsv
config.tsv is going to be where all the default string constants are going to be stored. This includes such string constants as button names, default fare field names and so on. It may even eventually contain some console config settings that we might be changing later under the 'Config' primary button at the main menu.

* All C++ header files are now located in merely one header file, main.h.
My experience as far as header files goes says this is an efficient idea, however I do have belief this may lead to issues.

* The Draw_Box function has been reallocated to Utilities.CPP as there's reusability of this function in both allocate_fare and main_menu class